requirements.txt
dwave-ocean-sdk==1.2.0
scipy==1.1.0
pandas==0.23.2
numpy==1.15.0
